// web/src/types/dashboard.ts
export interface DashboardTotalsProps {
  total_users: number
  total_teams: number
  total_organizations: number
  total_credentials: number
  total_inventories: number
  total_groups: number
  total_hosts: number
  total_projects: number
  total_templates: number
  total_schedules: number
  total_jobs: number
}